Add some ecs logic

+++ Assets/CodeBase/Infrastructure/StaticData/SpawnerPreferences.cs
@@ -5,7 +5,7 @@ namespace CodeBase.Infrastructure.StaticData
     [CreateAssetMenu(fileName = "SpawnerPreferences", menuName = "Preferences/Create Spawner Preferences")]
     public class SpawnerPreferences : ScriptableObject
     {
-        public int BubblesPerSecond;
+        public float BubblesPerSecond;
         public float MinSpeed;
         public float MaxSpeed;
         public float MinSize;
